requirements.txt
# Core dependencies
tensorflow>=2.14.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Development tools (optional)
pytest>=7.4.0  # For testing
black>=23.7.0  # For code formatting
flake8>=6.1.0  # For linting